+++ app/src/main/java/elder/ly/mobile/ui/components/SpecialtyList.kt
@@ -1,0 +1,27 @@
+package elder.ly.mobile.ui.components
+
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.PaddingValues
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.unit.dp
+import elder.ly.mobile.ui.screens.SpecialtyItem
+
+@Composable
+fun SpecialtyList(specialties: List<String>, onRemove: (String) -> Unit) {
+    LazyVerticalGrid(
+        columns = GridCells.Fixed(2),
+        contentPadding = PaddingValues(8.dp),
+        horizontalArrangement = Arrangement.spacedBy(8.dp),
+        verticalArrangement = Arrangement.spacedBy(8.dp)
+    ) {
+        items(specialties) { specialty ->
+            SpecialtyItem(
+                text = specialty,
+                onRemove = { onRemove(specialty) }
+            )
+        }
+    }
+}
