+++ app/src/main/java/elder/ly/mobile/ui/screens/PesquisaCuidador2.kt
@@ -1,4 +1,4 @@
-package elder.ly.mobile
+package elder.ly.mobile.ui.screens
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
@@ -18,7 +18,6 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.draw.shadow
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.input.pointer.motionEventSpy
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
@@ -26,8 +25,9 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import elder.ly.mobile.R.drawable.ic_launcher_background
 import elder.ly.mobile.R.drawable.ic_pesquisar
+import elder.ly.mobile.ui.components.Feature
+import elder.ly.mobile.ui.components.NavBar
 import elder.ly.mobile.ui.theme.MobileTheme
-import elder.ly.mobile.ui.theme.backgroundCustomBlue
 import elder.ly.mobile.ui.theme.secondaryContainerLight
 import elder.ly.mobile.ui.theme.tertiaryContainerLight
 
@@ -174,7 +174,7 @@ fun CardCuidador(){
 
 @Preview(showBackground = true)
 @Composable
-fun GreetingPreview() {
+fun PesquisaCuidadorPreview() {
     MobileTheme {
         PesquisaCuidador2()
     }
