+++ app/src/main/res/drawable/logo_google.xml
@@ -1,0 +1,26 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="2003dp"
+    android:height="2044dp"
+    android:viewportWidth="2003"
+    android:viewportHeight="2044">
+  <path
+      android:pathData="M980.5,-0.5C1008.2,-0.5 1035.8,-0.5 1063.5,-0.5C1310.5,10.2 1524.5,98.7 1705.5,265C1609.3,359.3 1513,453.5 1416.5,547.5C1291.4,435.2 1144.4,385 975.5,397C790.4,415.6 641.9,498.4 530,645.5C491.2,698.7 461,756.7 439.5,819.5C328.9,734.2 218.5,648.5 108.5,562.5C201.8,378.7 337.8,235.2 516.5,132C660.7,51.1 815.3,7 980.5,-0.5Z"
+      android:strokeAlpha="0.997"
+      android:fillColor="#eb4334"
+      android:fillAlpha="0.997"/>
+  <path
+      android:pathData="M108.5,562.5C218.5,648.5 328.9,734.2 439.5,819.5C395.2,954.3 395.5,1089 440.5,1223.5C330.4,1308.4 220.4,1393.6 110.5,1479C109.6,1479.5 108.6,1479.7 107.5,1479.5C42.7,1349 6.7,1211 -0.5,1065.5C-0.5,1036.2 -0.5,1006.8 -0.5,977.5C6.6,831.4 42.9,693.1 108.5,562.5Z"
+      android:strokeAlpha="0.998"
+      android:fillColor="#fabb05"
+      android:fillAlpha="0.998"/>
+  <path
+      android:pathData="M2002.5,1002.5C2002.5,1029.5 2002.5,1056.5 2002.5,1083.5C1998.8,1222.4 1970.3,1356.1 1917,1484.5C1865.8,1603.6 1793,1707.3 1698.5,1795.5C1590.8,1712.1 1483.1,1628.8 1375.5,1545.5C1462.2,1483.5 1524.4,1402.5 1562,1302.5C1572.6,1274 1579.8,1244.7 1583.5,1214.5C1395.8,1214.5 1208.2,1214.5 1020.5,1214.5C1020.5,1087.8 1020.5,961.2 1020.5,834.5C1340.8,834.5 1661.2,834.5 1981.5,834.5C1993.6,890 2000.6,946 2002.5,1002.5Z"
+      android:strokeAlpha="0.998"
+      android:fillColor="#4285f3"
+      android:fillAlpha="0.998"/>
+  <path
+      android:pathData="M440.5,1223.5C488.8,1363.3 574.8,1473.5 698.5,1554C826.2,1632.3 964.6,1661.6 1113.5,1642C1209,1631 1296.3,1598.8 1375.5,1545.5C1483.1,1628.8 1590.8,1712.1 1698.5,1795.5C1593.1,1892.1 1471.1,1960.6 1332.5,2001C1242.8,2026.4 1151.4,2040.6 1058.5,2043.5C1034.8,2043.5 1011.2,2043.5 987.5,2043.5C729.7,2031.6 506.7,1938.4 318.5,1764C231.7,1681.2 161.4,1586.3 107.5,1479.5C108.6,1479.7 109.6,1479.5 110.5,1479C220.4,1393.6 330.4,1308.4 440.5,1223.5Z"
+      android:strokeAlpha="0.998"
+      android:fillColor="#34a753"
+      android:fillAlpha="0.998"/>
+</vector>
